User can now pass -mcmc 1e7 -mcsave 1e4

+++ admb/trunk/src/nh99/xxmcmc.cpp
@@ -514,7 +514,7 @@ void function_minimizer::mcmc_routine(int nmcmc,int iseed0,double dscale,
 
       if ( (on=option_match(ad_comm::argc,ad_comm::argv,"-mcsave"))>-1)
       {
-        int jj=atoi(ad_comm::argv[on+1]);
+        int jj=(int)atof(ad_comm::argv[on+1]);
         if (jj <=0)
         {
           cerr << " Invalid option following command line option -mcsave -- "
